Update Windows build to work with latest changes.
git-svn-id: https://svn.apache.org/repos/asf/activemq/activemq-cpp/trunk/activemq-cpp@774892 13f79535-47bb-0310-9956-ffa450edef68

+++ src/test/decaf/internal/nio/DoubleArrayBufferTest.cpp
@@ -509,7 +509,7 @@ void DoubleArrayBufferTest::testPutIndexed() {
 
     for( std::size_t i = 0; i < testBuffer1->capacity(); i++ ) {
         CPPUNIT_ASSERT( testBuffer1->position() == 0 );
-        DoubleBuffer& ret = testBuffer1->put(i, i );
+        DoubleBuffer& ret = testBuffer1->put( i, (double)i );
         CPPUNIT_ASSERT( testBuffer1->get(i) == i );
         CPPUNIT_ASSERT( &ret == testBuffer1 );
     }
